requirements.txt
discord.py
matplotlib
Pillow
python-dotenv
motor
rich